Generate rail_soc_empty project
The project is created from simplicity_sdk_2025.6.2 version.
The project is created with copy import mode to export the simplicity
SDK files which makes easier to humans to use VS code workspace
for all code in the project. This way the AI agents can search in SDK
code as well.
The SDK files are not committed because those can be generated
from SDK through Simplicity Studio or SLC CLI utility.
The .gitignore is extended to ignore autogenerated files.